// src/api/models/APINotification.ts
export type APINotification = {
  actor_first_name: string;
  actor_id: string;
  actor_last_name: string;
  created_on: string;
  entity_id: number;
  entity_type: number;
  feed_post_title: string;
  file_name: string;
  full_note_url: string;
  id: number;
  note_url: string;
  notification_text: string;
  post_id: number;
  post_type_id: number;
  deck_size: number;
  profile_image_url: string;
  state: number;
};